reference/competitors/Memori/demos/personal_diary_assistant/requirements.txt
# Personal Diary Assistant - Core Dependencies
memorisdk>=0.1.0
litellm>=1.0.0
python-dotenv>=1.0.0

# Streamlit UI Dependencies  
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Additional useful packages
openai>=1.0.0
requests>=2.31.0
